backend/requirements.txt
django==2.2.6
psycopg2==2.9.1
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
Pillow==8.4.0
drf-extra-fields==3.1.1
reportlab==3.6.2